+++ ascriptor/lib/ascriptor/base.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module Ascriptor
-  class Base < RootObject::Base; end
-end