+++ ascriptor/spec/ascriptor/base_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe Ascriptor::Base, type: :ascriptor do
-  subject { described_class }
-
-  it { is_expected.to inherit_from RootObject::Base }
-end